Alexandria 2.2

+++ html/web/admin/blocks.php
@@ -130,32 +130,19 @@
     if (!get_protected_agora()) {
         $strdelete = '';
     }
-    //************ FI
+    //************ FI    
     $table->define_headers(array($strname, $strcourses, $strversion, $strhide.'/'.$strshow, $strprotecthdr, $strdelete, $strsettings));
     $table->define_baseurl($CFG->wwwroot.'/'.$CFG->admin.'/blocks.php');
     $table->set_attribute('class', 'compatibleblockstable blockstable generaltable');
     $table->setup();
     $tablerows = array();
 
-    // Sort blocks using current locale.
-    $blocknames = array();
     foreach ($blocks as $blockid=>$block) {
-        $blockname = $block->name;
-        if (file_exists("$CFG->dirroot/blocks/$blockname/block_$blockname.php")) {
-            $blocknames[$blockid] = get_string('pluginname', 'block_'.$blockname);
-        } else {
-            $blocknames[$blockid] = $blockname;
-        }
-    }
-    collatorlib::asort($blocknames);
-
-    foreach ($blocknames as $blockid=>$strblockname) {
-        $block = $blocks[$blockid];
         $blockname = $block->name;
 
         if (!file_exists("$CFG->dirroot/blocks/$blockname/block_$blockname.php")) {
             $blockobject  = false;
-            $strblockname = '<span class="notifyproblem">'.$strblockname.' ('.get_string('missingfromdisk').')</span>';
+            $strblockname = '<span class="notifyproblem">'.$blockname.' ('.get_string('missingfromdisk').')</span>';
             $plugin = new stdClass();
             $plugin->version = $block->version;
 
@@ -170,9 +157,18 @@
                 $incompatible[] = $block;
                 continue;
             }
+            $strblockname = get_string('pluginname', 'block_'.$blockname);
         }
 
-        $delete = '<a href="blocks.php?delete='.$blockid.'&amp;sesskey='.sesskey().'">'.$strdelete.'</a>';
+        //XTEC ************ AFEGIT - To let access only to xtecadmin user
+        //2012.06.25  @sarjona
+        if (get_protected_agora()) {
+        //************ FI    
+            $delete = '<a href="blocks.php?delete='.$blockid.'&amp;sesskey='.sesskey().'">'.$strdelete.'</a>';
+        //XTEC ************ AFEGIT - To let access only to xtecadmin user
+        //2012.06.25  @sarjona
+        }
+        //************ FI        
 
         $settings = ''; // By default, no configuration
         if ($blockobject and $blockobject->has_config()) {
@@ -207,10 +203,10 @@
             $visible = '';
         } else if ($blocks[$blockid]->visible) {
             $visible = '<a href="blocks.php?hide='.$blockid.'&amp;sesskey='.sesskey().'" title="'.$strhide.'">'.
-                       '<img src="'.$OUTPUT->pix_url('t/hide') . '" class="iconsmall" alt="'.$strhide.'" /></a>';
+                       '<img src="'.$OUTPUT->pix_url('i/hide') . '" class="icon" alt="'.$strhide.'" /></a>';
         } else {
             $visible = '<a href="blocks.php?show='.$blockid.'&amp;sesskey='.sesskey().'" title="'.$strshow.'">'.
-                       '<img src="'.$OUTPUT->pix_url('t/show') . '" class="iconsmall" alt="'.$strshow.'" /></a>';
+                       '<img src="'.$OUTPUT->pix_url('i/show') . '" class="icon" alt="'.$strshow.'" /></a>';
             $class = ' class="dimmed_text"'; // Leading space required!
         }
 
@@ -225,10 +221,10 @@
             $undeletable = '';
         } else if (in_array($blockname, $undeletableblocktypes)) {
             $undeletable = '<a href="blocks.php?unprotect='.$blockid.'&amp;sesskey='.sesskey().'" title="'.$strunprotect.'">'.
-                       '<img src="'.$OUTPUT->pix_url('t/unlock') . '" class="iconsmall" alt="'.$strunprotect.'" /></a>';
+                       '<img src="'.$OUTPUT->pix_url('t/unlock') . '" class="icon" alt="'.$strunprotect.'" /></a>';
         } else {
             $undeletable = '<a href="blocks.php?protect='.$blockid.'&amp;sesskey='.sesskey().'" title="'.$strprotect.'">'.
-                       '<img src="'.$OUTPUT->pix_url('t/lock') . '" class="iconsmall" alt="'.$strprotect.'" /></a>';
+                       '<img src="'.$OUTPUT->pix_url('t/unlock_gray') . '" class="icon" alt="'.$strprotect.'" /></a>';
         }
 
         $row = array(
@@ -240,7 +236,12 @@
             $delete,
             $settings
         );
-        $table->add_data($row);
+        $tablerows[] = array(strip_tags($strblockname), $row); // first element will be used for sorting
+    }
+
+    collatorlib::asort($tablerows);
+    foreach ($tablerows as $row) {
+        $table->add_data($row[1]);
     }
 
     $table->print_html();
